requirements.txt
cffi==1.15.0
decorator==4.4.2
gmpy2==2.1.2
joblib==1.1.0
networkx==2.5.1
numpy
pandas
pycparser==2.21
Pympler==1.0.1
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==0.24.2
scipy
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0